% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcflow-package.R
\docType{package}
\name{fcflow-package}
\alias{fcflow}
\alias{fcflow-package}
\title{fcflow: Forecast Workflow}
\description{
Tools for UHERO forecast model data preparation, estimation, simulation, and visualization workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uhero.github.io/fcflow/}
}

}
\author{
\strong{Maintainer}: Peter Fuleky \email{fuleky@hawaii.edu}

}
\keyword{internal}
